+++ tools/auth/auth_test.go
@@ -205,7 +205,7 @@ func TestNewProviderByName(t *testing.T) {
 		t.Errorf("Expected nil, got error %v", err)
 	}
 	if _, ok := p.(*auth.Yandex); !ok {
-		t.Error("Expected to be instance of *auth.yandex")
+		t.Error("Expected to be instance of *auth.Yandex")
 	}
 
 	// patreon
@@ -215,5 +215,14 @@ func TestNewProviderByName(t *testing.T) {
 	}
 	if _, ok := p.(*auth.Patreon); !ok {
 		t.Error("Expected to be instance of *auth.Patreon")
+	}
+
+	// mailcow
+	p, err = auth.NewProviderByName(auth.NameMailcow)
+	if err != nil {
+		t.Errorf("Expected nil, got error %v", err)
+	}
+	if _, ok := p.(*auth.Mailcow); !ok {
+		t.Error("Expected to be instance of *auth.Mailcow")
 	}
 }