+++ ui/src/components/settings/providers/SelfHostedOptions.svelte
@@ -3,18 +3,22 @@
 
     export let key = "";
     export let config = {};
+    export let required = false;
+    export let title = "Provider endpoints";
+
+    $: isRequired = required && config?.enabled;
 </script>
 
-<div class="section-title">Selfhosted endpoints (optional)</div>
-<Field class="form-field" name="{key}.authUrl" let:uniqueId>
+<div class="section-title">{title}</div>
+<Field class="form-field {isRequired ? 'required' : ''}" name="{key}.authUrl" let:uniqueId>
     <label for={uniqueId}>Auth URL</label>
-    <input type="url" id={uniqueId} bind:value={config.authUrl} />
+    <input type="url" id={uniqueId} bind:value={config.authUrl} required={isRequired} />
 </Field>
-<Field class="form-field" name="{key}.tokenUrl" let:uniqueId>
+<Field class="form-field {isRequired ? 'required' : ''}" name="{key}.tokenUrl" let:uniqueId>
     <label for={uniqueId}>Token URL</label>
-    <input type="url" id={uniqueId} bind:value={config.tokenUrl} />
+    <input type="url" id={uniqueId} bind:value={config.tokenUrl} required={isRequired} />
 </Field>
-<Field class="form-field" name="{key}.userApiUrl" let:uniqueId>
+<Field class="form-field {isRequired ? 'required' : ''}" name="{key}.userApiUrl" let:uniqueId>
     <label for={uniqueId}>User API URL</label>
-    <input type="url" id={uniqueId} bind:value={config.userApiUrl} />
+    <input type="url" id={uniqueId} bind:value={config.userApiUrl} required={isRequired} />
 </Field>